requirements.txt
python-dotenv
httpx
rich
schedule
typer
qrcode
pillow
requests
beautifulsoup4
lxml
readability-lxml
markdownify
python-frontmatter
google-api-python-client
google-auth
google-auth-oauthlib
